Dispose output progress monitor
The task runner progress monitor creates an output progress monitor
from the LogView but was not disposing it. Still seeing a null reference
exception sometimes when tasks are run during a build.

Object reference not set to an instance of an object
  at MonoDevelop.Ide.Gui.Components.LogViewProgressMonitor.OnWriteLog (System.String message) [0x00000] in
main/src/core/MonoDevelop.Ide/MonoDevelop.Ide.Gui.Components/LogView.cs:863