Fix null reference when tasks are run during build/clean
Calling the progress monitor's Log.WriteLine is not safe to do
after the task process has finished. This can cause the progress
monitor to try to write to the log view pad when the log view
progress monitor has been disposed, since the task's process has
completed, and was causing a null reference exception. Now the
exit code information is sent directly to the LogView after the
task is finished instead of using the progress monitor.

Fixes #1