On toggling a task binding refresh just the one tree node
Previously the entire bindings tree was re-created when a binding
changed, causing the tree to be collapsed. Now just the tree node
associated with the binding is refreshed and expanded. Also if the
binding tree is not visible, due to a task having been run, then it
will now be made the active tab.