Fix same task runner config file displayed for solution and project
When the task runner config file is in the same directory as the
solution and project file then it would be available in the Task
Runner Explorer window for both the solution and the project. Now
it is only available for the project if the file exists for both
the solution and project. This mirrors the behaviour of Visual Studio
on Windows.

Fixes #4