Add Task Runners Bundle addin
Add an .addin.xml file and associated build script which will
package all the task runners into a single .mpack that can be
installed. The main MonoDevelop.TaskRunner addin is still required
to be installed separately but the other Cake, Grunt, Gulp and
TypeScript task runners are installed with a single TaskRunnersBundle
addin instead of separately.